video_platform/requirements.txt
aenum==3.1.15
ai-api-client-sdk==2.0.4
ai-core-sdk==2.1.1
# apturl==0.5.2
asgiref==3.8.1
astroid==3.2.2
awscli==1.32.117
bcrypt==3.2.0
blinker==1.4
boto3==1.34.117
botocore==1.34.117
# Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
cryptography==3.4.8
# cupshelpers==1.0
Cython==3.0.10
# dbus-python==1.2.18
# defer==1.0.6
defusedxml==0.8.0rc2
Django==4.2.13
dill==0.3.8
distlib==0.3.8
distro==1.7.0
# distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
# dj-rest-auth==6.0.0
django-allauth==0.54.0
django-appconf==1.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
# django-ffmpeg==0.1.7
# django-heroku==0.3.1
# django-imagekit==5.0.0
django-ipware==7.0.1
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-authemail==2.1.7
django-social-share==2.3.0
django-templated-mail==1.1.1
django-use-email-as-username==1.4.0
djangorestframework==3.15.1
# djangorestframework-jwt==1.11.0
djoser==2.2.2
docutils==0.16
drf-yasg==1.21.7
# duplicity==0.8.21
fasteners==0.14.1
filelock==3.13.4
future==0.18.2
# gbinder-python==1.1.2
google-auth==2.29.0
graphviz==0.20.3
gunicorn==22.0.0
h11==0.14.0
httplib2==0.20.2
# idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
isort==5.13.2
jeepney==0.7.1
jmespath==1.0.1
jwcrypto==1.5.6
# keyring==23.5.0
kubernetes==29.0.0
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
# louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.7.0
metaflow==2.8.0
monotonic==1.6
more-itertools==8.10.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
pilkit==3.0
Pillow==9.0.1
pipenv==2023.12.1
platformdirs==4.2.1
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
# pycups==2.0.1
pydot==2.0.0
Pygments==2.11.2
# PyGObject==3.42.1
pyhumps==3.8.0
# PyJWT==1.7.1
pylint==3.2.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyRFC3339==1.1
# python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
# python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-ipware==3.0.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
# PyYAML==5.4.1
# reportlab==3.6.8
# requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.10.1
sap-ai-core-metaflow==1.1.14
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.4.1
# social-auth-core==4.5.4
sqlparse==0.5.0
# systemd-python==234
tomli==2.0.1
tomlkit==0.12.5
# typing_extensions==4.11.0
# ubuntu-drivers-common==0.0.0
# ubuntu-pro-client==8001
# unattended-upgrades==0.1
uritemplate==4.1.1
# urllib3==1.26.5
# usb-creator==0.3.7
# uvicorn==0.30.0
# virtualenv==20.26.0
# wadllib==1.3.6
# websocket-client==1.8.0
whitenoise==6.6.0
# xdg==5
# xkit==0.0.0
# zipp==1.0.0
django-storages==1.14.3
moviepy==1.0.3
ffmpy==0.3.2
ffmpeg-python==0.2.0
zencoder==0.6.5
imageio==2.34.1
imageio-ffmpeg==0.5.1
Pillow==9.0.1
